# docs/api/openapi-draft.yaml
openapi: 3.1.0
info:
  title: Immune Repertoire Platform API
  version: 0.2.0
  description: >
    Stable API contract for the frontend/backend separation migration. The current
    implementation is served by Flask (bridge phase), while this document defines
    the stable project, asset, job, and result surface that the standalone frontend
    should consume.

    **Implementation status (2026-06-29):**
    - ✅ Implemented in Flask → see `flask_app/routes/api_projects.py`, `api_jobs.py`
    - 🔄 Bridge implementation (Flask forwarding to module routes)
    - ⬜ Planned, not yet implemented

    See `docs/migration-progress.md` for detailed phase tracking.
servers:
  - url: http://127.0.0.1:5000
    description: Local Flask API
  - url: http://127.0.0.1:5173
    description: Vite dev server proxy
...
paths:
...
  /api/projects:
    get:
      tags: [Projects]
      summary: List projects for the current user
      x-status: ✅
      parameters:
        - name: name
          in: query
          schema:
            type: string
        - name: institution
          in: query
          schema:
            type: string
        - name: cooperation_level
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Project list
          content:
            application/json:
              schema:
                type: object
                required: [projects]
                properties:
                  projects:
                    type: array
                    items:
                      $ref: "#/components/schemas/Project"
    post:
      tags: [Projects]
      summary: Create a project
      x-status: ✅
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectCreate"
      responses:
        "201":
          description: Created project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
  /api/projects/{project_id}:
    get:
      tags: [Projects]
      summary: Get project detail with preview assets and samples
      x-status: ✅
      parameters:
        - $ref: "#/components/parameters/ProjectId"
      responses:
        "200":
          description: Project detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectDetail"
    patch:
      tags: [Projects]
      summary: Update a project
      x-status: ✅
      parameters:
        - $ref: "#/components/parameters/ProjectId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectUpdate"
      responses:
        "200":
          description: Updated project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
...
  /api/projects/{project_id}/results:
    get:
      tags: [Results]
      summary: List reusable project analysis results
      x-status: ✅
      parameters:
        - $ref: "#/components/parameters/ProjectId"
        - name: analysis_type
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Project results
          content:
            application/json:
              schema:
                type: object
                required: [success, results]
                properties:
                  success:
                    type: boolean
                  results:
                    type: array
                    items:
                      $ref: "#/components/schemas/Asset"
...
components:
  parameters:
    ProjectId:
      name: project_id
      in: path
      required: true
      schema:
        type: string
...
  schemas:
    Project:
      type: object
      required: [id, name, status]
      properties:
        id:
          type: string
        name:
          type: string
        user_id:
          type: integer
          nullable: true
        institution:
          type: string
          nullable: true
        cooperation_level:
          type: string
          nullable: true
        description:
          type: string
          nullable: true
        status:
          type: string
        asset_counts:
          type: object
          additionalProperties:
            type: integer
        sample_count:
          type: integer
        result_count:
          type: integer
        group_spec_count:
          type: integer
        created_at:
          type: string
          format: date-time
          nullable: true
        updated_at:
          type: string
          format: date-time
          nullable: true
    ProjectDetail:
      allOf:
        - $ref: "#/components/schemas/Project"
        - type: object
          properties:
            assets:
              type: array
              items:
                $ref: "#/components/schemas/Asset"
            group_specs:
              type: array
              items:
                type: object
            samples_preview:
              type: array
              items:
                type: object
    ProjectCreate:
      type: object
      required: [name]
      properties:
        name:
          type: string
        institution:
          type: string
        cooperation_level:
          type: string
        description:
          type: string
        status:
          type: string
          default: active
    ProjectUpdate:
      type: object
      additionalProperties: false
      properties:
        name:
          type: string
        institution:
          type: string
        cooperation_level:
          type: string
        description:
          type: string
        status:
          type: string
    Asset:
      type: object
      required: [id, project_id, asset_type, original_name, storage_path, size]
      properties:
        id:
          type: string
        project_id:
          type: string
        asset_type:
          type: string
        original_name:
          type: string
        storage_path:
          type: string
          description: Legacy local path. Replace with storage_uri during Phase 4.
        storage_uri:
          type: string
          nullable: true
          description: Stable storage URI. During the Flask transition this is surfaced from metadata.storage_uri.
        mime_type:
          type: string
          nullable: true
        size:
          type: integer
        metadata:
          type: object
          additionalProperties: true
        uploaded_at:
          type: string
          format: date-time
          nullable: true
...